transformer_model/requirements.txt
sonopy==0.1.2
torchmetrics>=0.8.1
torchaudio